pyproject.toml
[project]
name = "morningstar-fund-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ruff>=0.11.10",
]

[dependency-groups]
dev = [
    "assertpy>=1.1",
    "black>=25.1.0",
    "pytest>=8.3.5",
    "python-dateutil>=2.9.0.post0",
]
